Fixes/fixed issues 134 135 136 139 (#147)
* Fixed users have to scroll when the list of services reaches the maximum page size

#136

* Fixed Address Groups List Breaks With Large Number of services or long service names

#134

* App Identity hidden in the settings page interface

#140

* Added Title bars name #139
Fixed user has to refresh after importing keys #135

* Removed unnecessary descriptions and utilities from ApplicationSettings actions and layouts. Enhanced description templating for consistency with `appName`.

* Update apps/provider/src/app/admin/(internal)/groups/page.tsx

---------
