+++ apps/provider/src/actions/ApplicationSettings.ts
@@ -13,6 +13,7 @@ import urlJoin from "url-join";
 import { getServerApolloClient } from '@igniter/ui/graphql/server'
 import { indexerStatusDocument } from '@igniter/graphql'
 import {env} from "@/config/env";
+import { revalidateTag, unstable_cache } from 'next/cache'
 
 const UrlSchema = z.string().url("Please enter a valid URL").min(1, "URL is required")
 
@@ -35,6 +36,26 @@ const CreateSettingsSchema = z.object({
   chainId: z.nativeEnum(ChainId),
   updatedAtHeight: z.string(),
 })
+
+const getAppName = unstable_cache(
+  async () => {
+    const appSettings = await fetchApplicationSettings();
+
+    return appSettings.name;
+  },
+  undefined,
+  {tags: ['appName']}
+)
+
+export async function GetAppName() {
+  let appName = await getAppName();
+
+  if (!appName) {
+    appName = await fetchApplicationSettings().then(appSettings => appSettings.name)
+  }
+
+  return appName || 'Stake Igniter Provider';
+}
 
 export async function GetApplicationSettings() {
   return await fetchApplicationSettings();
@@ -76,6 +97,8 @@ export async function UpsertApplicationSettings(
       updatedBy: userIdentity,
     });
   }
+
+  revalidateTag('appName')
 }
 
 export async function completeSetup() {
