+++ apps/provider/src/app/admin/setup/page.tsx
@@ -1,11 +1,20 @@
+import type { Metadata } from 'next'
 import { redirect } from "next/navigation";
 import { Stepper } from "./stepper";
 import { auth } from "@/auth";
 import {UserRole} from "@/db/schema";
-import { GetApplicationSettings } from "@/actions/ApplicationSettings";
+import { GetApplicationSettings, GetAppName } from "@/actions/ApplicationSettings";
 import React from "react";
 
 export const dynamic = "force-dynamic";
+
+export async function generateMetadata(): Promise<Metadata> {
+  const appName = await GetAppName()
+
+  return {
+    title: `Setup - ${appName}`,
+  }
+}
 
 export default async function Page() {
   const settings = await GetApplicationSettings();
